feat: refactor numeric chart widget and introduce mini chart component
- Added `showMiniChart` property to `WidgetAutopilotComponent` and `WidgetNumericComponent` to control mini chart visibility.
