+++ src/app/widgets/minichart/minichart.component.html
