+++ src/app/widgets/minichart/minichart.component.scss
