+++ src/app/widgets/widget-autopilot/widget-autopilot.component.ts
@@ -291,6 +291,7 @@ export class WidgetAutopilotComponent extends BaseWidgetComponent implements OnI
       showMin: false,
       numDecimal: 1,
       numInt: 1,
+      showMiniChart: false,
       color: this.embedWidgetColor,
       enableTimeout: false,
       dataTimeout: 5,
@@ -343,6 +344,7 @@ export class WidgetAutopilotComponent extends BaseWidgetComponent implements OnI
       showMin  : false,
       numDecimal  : 1,
       numInt  : 1,
+      showMiniChart: false,
       color  : this.embedWidgetColor,
       enableTimeout  : false,
       dataTimeout  : 5,
@@ -372,6 +374,7 @@ export class WidgetAutopilotComponent extends BaseWidgetComponent implements OnI
       showMin: false,
       numDecimal: 1,
       numInt: 1,
+      showMiniChart: false,
       color: this.embedWidgetColor,
       enableTimeout: false,
       dataTimeout: 5,
