--- src/app/widgets/widget-numeric-chart/widget-numeric-chart.component.html
@@ -1,13 +1,0 @@
-<widget-host [(config)]="widgetProperties.config" [id]="widgetProperties.uuid" (configChange)="updateConfig($event)">
-  <div class="textGenericWrapper" (onResize)="onResized($event)">
-    <widget-title [text]="widgetProperties.config.displayName" [color]="labelColor()"/>
-    @if (showMiniChart()) {
-      <simple-data-chart
-        [theme]="this.theme()"
-        />
-    }
-    <canvas #canvasUnit class="canvas"></canvas>
-    <canvas #canvasValue class="canvas"></canvas>
-    <canvas #canvasMinMax class="canvas"></canvas>
-  </div>
-</widget-host>
