--- src/app/widgets/widget-numeric-chart/widget-numeric-chart.component.scss
@@ -1,20 +1,0 @@
-:host {
-  display: block;
-  width: 100%;
-  height: 100%;
-}
-
-.textGenericWrapper {
-  position: absolute;
-  display: block;
-  width: 100%;
-  height: 100%;
-  z-index: inherit;
-}
-
-.canvas {
-  display: block;
-  position: absolute;
-  top: 0;
-  left: 0;
-}
