--- src/app/widgets/widget-numeric-chart/widget-numeric-chart.spec.ts
@@ -1,26 +1,0 @@
-
-import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
-
-import { WidgetNumericChartComponent } from './widget-numeric-chart.component';
-
-describe('WidgetNumericChartComponent', () => {
-  let component: WidgetNumericChartComponent;
-  let fixture: ComponentFixture<WidgetNumericChartComponent>;
-
-  beforeEach(waitForAsync(() => {
-    TestBed.configureTestingModule({
-    imports: [WidgetNumericChartComponent]
-})
-    .compileComponents();
-  }));
-
-  beforeEach(() => {
-    fixture = TestBed.createComponent(WidgetNumericChartComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should be created', () => {
-    expect(component).toBeTruthy();
-  });
-});
