+++ src/app/widgets/widget-numeric/widget-numeric.component.html
@@ -1,6 +1,11 @@
 <widget-host [(config)]="widgetProperties.config" [id]="widgetProperties.uuid" (configChange)="updateConfig($event)">
   <div class="textGenericWrapper" (onResize)="onResized($event)">
     <widget-title [text]="widgetProperties.config.displayName" [color]="labelColor()"/>
+    @if (showMiniChart()) {
+      <minichart
+        [theme]="this.theme()"
+        />
+    }
     <canvas #canvasUnit class="canvas"></canvas>
     <canvas #canvasValue class="canvas"></canvas>
     <canvas #canvasMinMax class="canvas"></canvas>
